Update to use native container command
Update the build to use the inbuilt docker container command with the maven verify command to remove reliance on manually creating an Amazon Linux 2 container with required dependencies. We need to verify if we need the exact OS when creating the Native image but this can easily be restored if needed.