TXMA-147 Add SAM Template for audit with FireHose Service
Template added that creates a FireHose Stream that batches messages to an S3 bucket.

Updated event processor template to integrate with the audit stack in order to allow messages to move between SNS and FireHose (We still need to make this deploy between accounts. Currently, this is between stacks only).

Shell script added to help run both scripts. This is likely to change as we look towards the deploy to build story.