TXMA-147 Add Cross account functionality between Event processing and Audit accounts.
I have updated the files to create the two stacks on separate accounts. This has been tested end to end and confirmed.

The configuration has been updated to run as much as possible from environment defined config in the toml files.

We now need to run the sam commands in run-sam-deploy.sh separately as we have a dependency between the two. We have options to get around this (Secrets manager or a Lambda Lookup for example) but this has been left with the dev platform team for now.