Update template to include resources specific to KBV, IPV and SPOT
update to create role specific to each lambda that can be shared with the external accounts used by the other teams when setting access to their SQS queues.

Put in a staging or greater source mapping to set up the event trigger between the queue and our Lambda functions with a condition not to run in dev or build.