Revert the hardcoded function name change. This was unnecessary as it was the EventSourceArn causing the issue with deployment and the reference the resource directly is preferable.
Having tested on the Dev environment I believe the issue is due to the YAML indentation of the EventSourceArn value. This instead needs to be after the colon and not tabbed underneath (I believe it tries to interpret it as a list which causes the issue with it not being recognised).