TMA-198 Fix issues in code with validation of unknown fields. This was causing testing on AWS to fail.
Change user message to have transcation_id instead of id field. This has changed in the ADR/Data dictionary since code was implemented

Remove need to use KeyValuePairs in extensible fields. This was a hard requirement due to previously using protocol buffers. However, we are now using JSON and we can simplify this to use the unknow type (as we will not know until runtime the fields supplied).