Make the component_id be the unique variable for each test. And fix a bug
The event_name needs to be used in the filter, so it cannot be amended in the code to be unique. The component_id, however, can be changed.

Also, to fix a bug, where either the cloudwatch logs were being called too many times a second, or too quickly so some of the logs were being missed, the sleep times have been edited.