Change purpose of a function
Instead of having a function to search for a new key and output its contents, we have a function which searches for the newest key and outputs its contents. This allows it to be called multiple times in case another key is produced before the one we wanted, but after the test started running.

To implement this, a loop is not in the step function which will run every 10 seconds which runs the function to find the latest objects. This loop will break if the correct object is found, it is the first time searching for an object and it has taken more than 100 seconds, or if it is run for another firehose (as this can only run if enough time has passed for the firehose to process the message).