Change addUniqueComponentID to addUniqueFields
Because not all subscribers to the SNS receive the component ID (i.e. the performance endpoint) it makes more sense to use the timestamp field as the unique field. To make sure it is unique, we create it to the nearest millisecond.

So timestamp is the current time, and component_id is still being added to match the account.